Support case-insensitive on non-Windows OSs.
The environment variable EnableGoCliScan existence gates usage of the Go
CLI tools for determing what modules are in-use.  The current check does
a get of the environment variable, and if it exists behavior is enabled.
On Windows this is case-insensitive, but on Linux (or MacOS) this is
case-sensitive so the user must exactly use the casing of
'EnableGoCliScan'.

Our CI system automatically capitalizes all environment variables when
they are defined, so EnableGoCliScan becomes ENABLEGOCLISCAN.  I am not
aware of a way to control this behavior, so there is no way to enable Go
CLI tooling.  My fix is to treat all environment variable exitence
checks as case-insensitive.