refactor(linux): use a smaller image to test base image annotations
Currently we're using an Ubuntu image, which is >100MB for unit tests. This is a large image for build machines and users to pull in for a single unit test, especially when we're only checking annotations.

This PR instead uses the `docker.io/library/hello-world:latest` image, which weighs in at a whopping 13KB.