chore: codestyle and roslyn analyzers best practices
Currently, we've been using [FxCop][1] to run code analysis and enforce some best practices. FxCop has been deprecated since around 2019. Similar code analysis has been ported to the .NET Roslyn Analyzers framework, in the `Microsoft.CodeAnalysis.FxCopAnalyzers` package, however that package is no longer required as of .NET 5.0 as the analyzers are included in the .NET SDK itself[^1]. As we are planning to upgrade to .NET 6.0 soon[^2] I am removing all reference to FxCop.

This change adds the `.editorconfig` from the .NET Boxed template[^3] that we have used as a reference for other parts of our project layout. I've not vetted every single rule that it enables, but it provides good defaults and is trusted and used widely.

Other changes to MSBuild configuration include:

- [`EnforceCodeStyleInBuild`][2] set to `true`
- [`AnalysisLevel`][3] set to `latest`
- [`AnalysisMode`][4] set to `AllEnabledByDefault`
- [`TreatWarningsAsErrors`] set to `false` (so these new code analysis changes wont break builds yet)

I plan to resolve some of the errors in followup PRs, but wanted to minimize the size of this one to enable easier review.

[1]: https://docs.microsoft.com/en-us/visualstudio/code-quality/net-analyzers-faq?view=vs-2022
[2]: https://docs.microsoft.com/en-us/dotnet/core/project-sdk/msbuild-props#enforcecodestyleinbuild
[3]: https://docs.microsoft.com/en-us/dotnet/core/project-sdk/msbuild-props#analysislevel
[4]: https://docs.microsoft.com/en-us/dotnet/core/project-sdk/msbuild-props#analysismode

[^1]: https://docs.microsoft.com/en-us/visualstudio/code-quality/migrate-from-fxcop-analyzers-to-net-analyzers?view=vs-2022
[^2]: https://github.com/microsoft/component-detection/issues/107
[^3]: https://github.com/Dotnet-Boxed/Templates/tree/main/Source/NuGetTemplate