test: exclude `Microsoft.NET.Test.Sdk` from `Microsoft.ComponentDetection.TestsUtilities`
The `Microsoft.NET.Test.Sdk` automatically creates an entrypoint for test projects and enables them to run. The `Microsoft.ComponentDetection.TestsUtilities` project is a shared library that is used in other test projects, and is not a runnable project itself.

This change moves `Microsoft.NET.Test.Sdk` from `test/Directory.Build.props`, which is inherited by all projects in the `test` directory, to each individual test project that needs it.