Add DotNetComponent
The DotNetComponent type is used to report usage of the .NET SDK version

This is important as the .NET SDK will add redistributable content to
the project, so using an older version may mean your application
contains unserviced copies of the host or runtime.
The TargetFramework and project type is reported as well since these
influence what content the SDK may reference.