Fix dotnet detector for exes / newer version not on path
We weren't handling EXEs correctly - simply missed this case.

When a global.json mentions a version that doesn't exist on the path,
the `dotnet --version` command will fail, and the `commandLineInvocationService`
will throw InvalidOperationException since it cannot distinguish this
failure from a missing command.  Catch this exception and treat it like
a failure so we fallback to parsing `global.json`.