Support moved project.assets.json files
We've seen cases where the project.assets.json files refer to paths with
a different root than the current scan.  One example is building inside
a container then running on the filesystem mounted outside the container.

Address this by comparing the relative path to the root, and finding if
the same relative path is present in the assets file's output path.