Fix dotnet detector for exes / newer version not on path (#1367)
* Fix dotnet detector for exes / newer version not on path

We weren't handling EXEs correctly - simply missed this case.

When a global.json mentions a version that doesn't exist on the path,
the `dotnet --version` command will fail, and the `commandLineInvocationService`
will throw InvalidOperationException since it cannot distinguish this
failure from a missing command.  Catch this exception and treat it like
a failure so we fallback to parsing `global.json`.

* Support moved project.assets.json files

We've seen cases where the project.assets.json files refer to paths with
a different root than the current scan.  One example is building inside
a container then running on the filesystem mounted outside the container.

Address this by comparing the relative path to the root, and finding if
the same relative path is present in the assets file's output path.